useChannelsWithNotification
